+++ src/Symfony/Component/HttpKernel/Tests/EventListener/TestSessionListenerTest.php
@@ -123,7 +123,7 @@ class TestSessionListenerTest extends TestCase
 
         $response = $this->filterResponse(new Request(), HttpKernelInterface::MASTER_REQUEST, $response);
 
-        $this->assertSame($expected, $response->headers->get('Set-Cookie', null, false));
+        $this->assertSame($expected, $response->headers->all()['set-cookie']);
     }
 
     public function anotherCookieProvider()